// src/5-entities/task/model/types.ts
export type Task = {
  id: string;
  title: string;
  description?: string;
  authorId: string;
  boardId?: string;
};

export type CreateTaskData = {
  title: string;
  description?: string;
  authorId: string;
  boardId?: string;
};

export type UpdateTaskData = {
  title?: string;
  description?: string;
  authorId?: string;
  boardId?: string;
};